+++ CHANGES.rst
@@ -7,8 +7,9 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
 4.1b4 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Adapted 'meetingadvicefinances_workflow' to use MeetingAdviceCommunesWorkflowActions/MeetingAdviceCommunesWorkflowConditions
+  instead the '@@advice-wf-conditions' view
+- Added workflow adaptation for the meetingadvicefinances_workflow to add the 'advicecreated' intial state
 
 4.1b3 (2019-05-16)
 ------------------